Design mdb Ecriture et lecture du délai d'impulsion de la machine OK Lecture du délai restant OK

+++ SW/Main/Satellite_3/include/main.h
@@ -24,6 +24,6 @@
 #include "parameters.h"
 #include "adc.h"
 
-int delayBlink;
+uint8_t delayBlink;
 
 #endif
